requirements.txt
aniso8601==3.0.2
certifi==2018.8.24
click==6.7
Django==2.0.7
Flask==1.0.2
Flask-RESTful==0.3.6
gensim==3.5.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.15.1
pipenv==2018.7.1
pymongo==3.4.0
pytz==2018.5
scipy==1.1.0
simple-intent-parser==1.5.14
six==1.11.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
